+++ R/myseq_n.R
@@ -12,7 +12,6 @@
 #'myseq_n(x = c(2, 4, 3), n = 5)
 #'myseq_n(x = c(2, 4, 3), n = 6)
 #'myseq_n(x = c(2, 4, 3), n = 7)
-#'
 myseq_n <- function(x, n){
   stopifnot(length(x) == 3) # error check x
   stopifnot(n > 0) # error check n